requirements.txt
# Core ML Libraries
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2+cu121
accelerate==1.3.0
transformers==4.48.2
bitsandbytes==0.45.1
sentencepiece==0.1.99

# Vision and Image Processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-image==0.25.0
pycocotools==2.0.8

# Data Processing and Utilities
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1
matplotlib==3.10.0
wandb==0.19.4
netifaces==0.11.0

# LLaVA and Dependencies
llava @ git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd
peft==0.4.0

# System and Performance
psutil==6.1.1
packaging==24.2
triton==2.1.0

# CUDA Dependencies
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26 